Removed RasterArgs interface
