+++ src/main/java/org/reactome/server/tools/diagram/exporter/raster/api/RasterArgs.java
@@ -1,72 +1,146 @@
 package org.reactome.server.tools.diagram.exporter.raster.api;
 
-
 import org.reactome.server.tools.diagram.exporter.raster.profiles.ColorProfiles;
 
 import java.awt.*;
+import java.util.Collection;
+import java.util.HashSet;
 import java.util.Set;
 
-/** Arguments for the diagram exporter */
-public interface RasterArgs {
+public class RasterArgs implements RasterArgs {
+
+	private String stId;
+	private Double factor = 1.;
+	private String format;
+	private String token;
+	private Set<String> flags;
+	private Set<String> selected;
+	private ColorProfiles profiles;
+	private Color background;
+	private Integer column;
+	private String resource;
+	private Boolean writeTitle;
+
+	public RasterArgs(String stId, String format) {
+		this.stId = stId;
+		setFormat(format);
+	}
+
+	/** diagram stable identifier */
+	public String getStId() {
+		return stId;
+	}
+
+	public void setStId(String stId) {
+		this.stId = stId;
+	}
+
+	/** quality of output image. number of pixels per point in the diagram */
+	public Double getFactor() {
+		return factor;
+	}
+
+	public void setQuality(Integer quality) {
+		if (quality != null)
+			this.factor = scale(quality);
+	}
+
+	private double scale(int factor) {
+		if (factor < 1 || factor > 10)
+			throw new IllegalArgumentException("factor must be in the range [1-10]");
+		if (factor < 5) {
+			return interpolate(factor, 1, 5, 0.1, 1);
+		} else return interpolate(factor, 5, 10, 1, 3);
+	}
+
+	private double interpolate(double x, double min, double max, double dest_min, double dest_max) {
+		return (x - min) / (max - min) * (dest_max - dest_min) + dest_min;
+	}
+
+	/** output image format (png, jpg, gif) */
+	public String getFormat() {
+		return format;
+	}
+
+	public void setFormat(String format) {
+		this.format = format == null ? "png" : format.trim().toLowerCase();
+	}
+
+	/** color profiles for diagram, analysis and interactors */
+	public ColorProfiles getProfiles() {
+		if (profiles == null)
+			profiles = new ColorProfiles(null, null, null);
+		return profiles;
+	}
+
+	public void setProfiles(ColorProfiles profiles) {
+		this.profiles = profiles;
+	}
+
+	public String getToken() {
+		return token;
+	}
+
+	/** Analysis token */
+	public void setToken(String token) {
+		this.token = token;
+	}
+
+	/** Background color when no transparency is available */
+	public Color getBackground() {
+		return background;
+	}
+
+	public void setBackground(Color color) {
+		this.background = color;
+	}
 
 	/**
-	 * @return diagram stable identifier
+	 * In case an expression analysis is run, the column to show. Leave it null
+	 * to generate a GIF with all the columns. May take longer.
 	 */
-	String getStId();
+	public Integer getColumn() {
+		return column;
+	}
 
-	/**
-	 * @return factor or scale. Resulting diagram dimension will be <em>(width *
-	 * factor, height * factor)</em>
-	 */
-	Double getFactor();
+	public void setColumn(Integer column) {
+		this.column = column;
+	}
 
-	/**
-	 * @return output image format: PNG, JPG, JPEG or GIF. Case insensitive.
-	 */
-	String getFormat();
+	@Override
+	public String getResource() {
+		return resource;
+	}
 
-	/**
-	 * @return analysis token. EXPRESSION or ENRICHMENT supported.
-	 */
-	String getToken();
+	public void setResource(String resource) {
+		this.resource = resource;
+	}
 
-	/**
-	 * @return list of elements to be flagged: stIds, dbIds, identifiers or gene
-	 * names.
-	 */
-	Set<String> getFlags();
+	@Override
+	public Boolean getWriteTitle() {
+		return writeTitle;
+	}
 
-	/**
-	 * @return list of elements to be selected: stIds, dbIds, identifiers or
-	 * gene names.
-	 */
-	Set<String> getSelected();
+	public void setWriteTitle(Boolean writeTitle) {
+		this.writeTitle = writeTitle;
+	}
 
-	/**
-	 * Names of the color profiles to use.
-	 * <p>
-	 * <table> <tr> <th>profile</th> <th>values</th> <th>default</th> </tr> <tr>
-	 * <td>diagram</td> <td>modern, standard</td> <td>modern</td> </tr> <tr>
-	 * <td>analysis</td> <td>standard, copper plus, strosobar</td>
-	 * <td>standard</td> </tr> <tr> <td>interactors</td> <td>cyan, teal</td>
-	 * <td>cyan</td> </tr> </table>
-	 *
-	 * @return color profiles (diagram, analysis and interactors)
-	 */
-	ColorProfiles getProfiles();
+	public Set<String> getFlags() {
+		return flags;
+	}
 
-	/**
-	 * @return for non transparent image formats, color of the background.
-	 */
-	Color getBackground();
+	public void setFlags(Collection<String> flags) {
+		if (flags != null)
+			this.flags = new HashSet<>(flags);
+	}
 
-	/**
-	 * @return if there is an expression analysis in the token, the column where
-	 * to take expression values (0 is the first). If null, first is used (0).
-	 * If <strong>format</strong> is <em>gif</em> and <strong>column</strong> is
-	 * <em>null</em> then the result will be an animated gif.
-	 */
-	Integer getColumn();
+	public Set<String> getSelected() {
+		return selected;
+	}
 
-	String getResource();
+	public void setSelected(Collection<String> selected) {
+		if (selected != null)
+			this.selected = new HashSet<>(selected);
+	}
+
 }
